Move tests next to production code
